# [exporter/splunkhec] asynchronous HTTP request consumption