@atoulme 2024-07-26
this is a costly issue. We might need to review a bit more the code and come up with a different approach.